+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Fix nullability annotation on `EntityChangeEmitter.attributeChanged#value`
+
 ### [v6.132](https://github.com/replicant4j/replicant/tree/v6.132) (2023-01-25) · [Full Changelog](https://github.com/spritz/spritz/compare/v6.131...v6.132)
 
 Changes in this release: